refactor(permissions): adresse les retours de review PR 2311
- Renomme permission/utils.ts → permission/order.ts et y centralise
  INDICATEUR_ACTION_ORDER / COLLECTION_ACTION_ORDER (supprime les
  doublons dans listerMesPermissions et loadPrincipalPermissions)
- Renomme *_READ_PERMISSIONS → *_PERMISSIONS_GRANTING_READ pour
  expliciter que ces listes accordent l'accès en lecture, pas qu'elles
  sont des permissions READ
- Renomme hasWriteData → hasIndicateurWriteData (cohérence de nommage)
- Fait importer CollectionPermissionAction / IndicateurPermissionAction
  depuis l'enum Prisma dans seed.ts (source de vérité directe)
