--- apps/kpilote-api/src/permission/utils.ts
@@ -1,2 +1,0 @@
-export const sortByOrder = <T>(items: T[], order: readonly T[]): T[] =>
-  [...items].sort((a, b) => order.indexOf(a) - order.indexOf(b))
